requirements.txt
langchain-core
langchain
langchain-groq
langgraph
python-dotenv
beautifulsoup4